Set max datepicker date

+++ app/src/main/java/com/utn/nutricionista/models/Weight.kt
@@ -2,7 +2,7 @@ package com.utn.nutricionista.models
 
 class Weight(var id:  String?,
              var uid : String? = null,
-             var peso: Double,
+             var peso: Float,
              var fecha : String)
 {
     fun date(): LocalDate {
